+++ skema/program_analysis/python2cast.py
@@ -4,7 +4,7 @@ import ast
 from skema.program_analysis import astpp
 import json
 import argparse
-from skema.program_analysis.PyAST2CAST import py_ast_to_cast
+from skema.program_analysis.CAST.pythonAST import py_ast_to_cast
 from skema.program_analysis.CAST2FN import cast
 from skema.program_analysis.CAST2FN.model.cast import SourceRef
 from skema.program_analysis.CAST2FN.cast import CAST
