+++ skema/program_analysis/tests/test_closures.py
@@ -4,7 +4,7 @@ from skema.program_analysis.multi_file_ingester import process_file_system
 from skema.gromet.fn import GrometFNModuleCollection
 import ast
 
-from skema.program_analysis.PyAST2CAST import py_ast_to_cast
+from skema.program_analysis.CAST.pythonAST import py_ast_to_cast
 from skema.program_analysis.CAST2FN.model.cast import SourceRef
 from skema.program_analysis.CAST2FN import cast
 from skema.program_analysis.CAST2FN.cast import CAST
