+++ skema/utils/script_functions.py
@@ -10,7 +10,7 @@ from skema.gromet.fn import (
 )
 
 from skema.utils.fold import dictionary_to_gromet_json, del_nulls
-from skema.program_analysis.PyAST2CAST import py_ast_to_cast
+from skema.program_analysis.CAST.pythonAST import py_ast_to_cast
 from skema.program_analysis.CAST2FN import cast
 from skema.program_analysis.CAST2FN.model.cast import SourceRef
 from skema.program_analysis.CAST2FN.cast import CAST
